feat(order-service): update-type-in-entity-order-item

+++ order-service/src/main/java/com/ecommerce/order_service/entity/OrderItem.java
@@ -32,6 +32,6 @@ public class OrderItem {
     Order order;
 
     int quantity;
-    BigDecimal price;
-    BigDecimal subTotal;
+    int price;
+    int subTotal;
 }
